+++ src/client/java/com/github/prohect/alias/builtinAlias/OpenInventoryAlias.java
@@ -4,7 +4,6 @@ import com.github.prohect.alias.Alias;
 import com.github.prohect.alias.BuiltinAliasWithBooleanArgs;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.gui.screen.Screen;
-import net.minecraft.client.gui.screen.ingame.CreativeInventoryScreen;
 import net.minecraft.client.gui.screen.ingame.InventoryScreen;
 
 public class OpenInventoryAlias
@@ -23,18 +22,18 @@ public class OpenInventoryAlias
     public OpenInventoryAlias run(String args) {
         parseArgs(args);
         // cancel open event from text input screen
-        if (Alias.isUnderTextInputScreen.get() && flag) return this;
+        if (Alias.isUnderTextInputScreen() && flag) return this;
         MinecraftClient mc = MinecraftClient.getInstance();
         if (mc.player == null) return this;
         if (flag) {
-            if (!Alias.isUnderAnyScreen.get()) {
+            if (!Alias.isUnderAnyScreen()) {
                 mc.setScreen(new InventoryScreen(mc.player));
             }
         } else {
-            if (Alias.isUnderAnyScreen.get()) {
-                Screen currentScreen = mc.currentScreen;
-                if (currentScreen instanceof InventoryScreen || currentScreen instanceof CreativeInventoryScreen) {
-                                    currentScreen.close();
+            if (Alias.isUnderAnyScreen()) {
+                Screen currentScreen = Alias.getCurrentScreen();
+                if (Alias.isInInventoryScreen() || Alias.isInCreativeInventoryScreen()) {
+                    currentScreen.close();
                 }
             }
         }