+++ src/client/java/com/github/prohect/alias/builtinAlias/RightAlias.java
@@ -15,7 +15,7 @@ public class RightAlias extends BuiltinAliasWithBooleanArgs<RightAlias> {
     public RightAlias run(String args) {
         parseArgs(args);
         // cancle press event from text input screen
-        if (Alias.isUnderTextInputScreen.get() && flag) return this;
+        if (Alias.isUnderTextInputScreen() && flag) return this;
         KeyBinding rightKey = MinecraftClient.getInstance().options.rightKey;
         rightKey.setPressed(flag);
         if (flag) rightKey.timesPressed++;