+++ src/client/java/com/github/prohect/alias/builtinAlias/SneakAlias.java
@@ -15,7 +15,7 @@ public class SneakAlias extends BuiltinAliasWithBooleanArgs<SneakAlias> {
     public SneakAlias run(String args) {
         parseArgs(args);
         // cancle press event from text input screen
-        if (Alias.isUnderTextInputScreen.get() && flag) return this;
+        if (Alias.isUnderTextInputScreen() && flag) return this;
         KeyBinding sneakKey = MinecraftClient.getInstance().options.sneakKey;
         sneakKey.setPressed(flag);
         if (flag) sneakKey.timesPressed++;